% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_detection.R
\name{detectInterestPoints}
\alias{detectInterestPoints}
\title{Detect multi-scale interest points using blob detection}
\usage{
detectInterestPoints(
  spe,
  assay.type = assayNames(spe)[1],
  masks = NULL,
  features = TRUE,
  thresh = 0,
  ipassay = "ipoints",
  octaves = NULL,
  BPPARAM = BiocParallel::bpparam(),
  use.dimred = NULL
)
}
\arguments{
\item{spe}{a SpatialExperiment object, containing gene/protein or any such omic measurement. NOTE that spatial coordinates need to be integers. This can be achieved by binning data into coarser bins. Binning will speed up computation without significantly affecting interest point detection and is therefore advised.}

\item{assay.type}{a character or numeric, specifying the assay to use (default is the first assay).}

\item{masks}{a named list of binary or logical matrices, that specifies the mask to use for each sample. The defauls (NULL) is to use mask out areas without any expression measurement. Otherwise, the list must be named using the unique \link{sample_id}'s and should cover all loci (i.e., nrow should be equal to the maximum y coordinate and ncol should equal the maximum x coordinate).}

\item{features}{a character or logical vector, specifying the features to be used. If character, these should be rownames of the SpatialExperiment object or the name of a column in colData(spe) containing logical values.}

\item{thresh}{a numeric, indicating the response cutoff for each interest point.}

\item{ipassay}{a character, specifying the prefix of the assay where detected interest points are stored.}

\item{octaves}{a numeric, stating the maximum number of octaves that should be computed (default NULL, max octave is automatically computed for each sample).}

\item{BPPARAM}{an optional \link{BiocParallelParam} instance determining the parallel back-end to be used during evaluation.}

\item{use.dimred}{a character or numeric, specifying whether existing values in \code{reducedDims(x)} should be used.}
}
\value{
a SpatialExperiment object containing an additional assay storing the interest points identified at each location. This assay is stored as a BumpyMatrix where each cell of the matrix (each location, gene/protein pair) stores a dataframe with 4 columns, scale, x, y, response. Each coordinate (x,y) represents the sub-pixel resolution of the interest point.
}
\description{
Treating spatial omics datasets as multi-channel images, this function builds a Gaussian scale-space pyramid and uses it to infer the scale of each gene at each locus. It then identifies local maxima. Local maxima above a threshold are points of interest that can be used to estimate local scales using the \code{\link[=detectScale]{detectScale()}} function. The function uses a fast approximation adapted from the SURF (speeded-up robust features) method.
}
\examples{
# ADD_EXAMPLES_HERE

}
