% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_detection.R
\name{detectScale}
\alias{detectScale}
\title{Estimate local length-scales using interest points}
\usage{
detectScale(
  spe,
  ipassay,
  masks = NULL,
  smooth = TRUE,
  thresh = 0,
  BPPARAM = BiocParallel::bpparam(),
  slot.type = c("auto", "assay", "dimred")
)
}
\arguments{
\item{spe}{a SpatialExperiment object, containing gene/protein or any such omic measurement. NOTE that spatial coordinates need to be integers. This can be achieved by binning data into coarser bins. Binning will speed up computation without significantly affecting interest point detection and is therefore advised.}

\item{ipassay}{a character, specifying the prefix of the assay where detected interest points are stored.}

\item{masks}{a named list of binary or logical matrices, that specifies the mask to use for each sample. The defauls (NULL) is to use mask out areas without any expression measurement. Otherwise, the list must be named using the unique \link{sample_id}'s and should cover all loci (i.e., nrow should be equal to the maximum y coordinate and ncol should equal the maximum x coordinate).}

\item{thresh}{a numeric, indicating the response cutoff for each interest point.}

\item{BPPARAM}{an optional \link{BiocParallelParam} instance determining the parallel back-end to be used during evaluation.}

\item{slot.type}{a character, specifying where interest points are saved. If expression data from an assay was used, this will be "assay", and if a reduced dimension was used, this will be "dimred". Leaving it at "auto" (the default) will enable a search across both and locate the assay. If assays with the name 'ipassay' exist in both 'assay' and 'dimred', users must specify as "auto" will throw an error.}
}
\value{
a numeric vector containing the estimated scales.
}
\description{
This function estimates the local length-scale for each location by combining scale information from interest points detected using each gene.
}
\examples{
# ADD_EXAMPLES_HERE

}
